+++ hooks/bench_max_rss.sh
@@ -1,18 +1,10 @@
 # Max resident set size
-rm -rf .swc/
 /usr/bin/time -v -o time_baseline.txt node runtime.mjs
-/usr/bin/time -v -o time_sync_babel.txt node --import ./sync-hooks-babel.mjs runtime.mjs
-/usr/bin/time -v -o time_sync_oxc.txt node --import ./sync-hooks-oxc.mjs runtime.mjs
-/usr/bin/time -v -o time_sync_oxc_wasm.txt node --import ./sync-hooks-oxc-wasm.mjs runtime.mjs
-/usr/bin/time -v -o time_sync_swc.txt node --import ./sync-hooks-swc.mjs runtime.mjs
-/usr/bin/time -v -o time_sync_acorn.txt node --import ./sync-hooks-acorn.mjs runtime.mjs
+/usr/bin/time -v -o time_sync_noop.txt node --import ./sync-hooks-noop.mjs runtime.mjs
+/usr/bin/time -v -o time_register_tap_oxc.txt node --import ./register-tap-oxc.mjs runtime.mjs
+/usr/bin/time -v -o time_register_tap_acorn.txt node --import ./register-tap-acorn.mjs runtime.mjs
 /usr/bin/time -v -o time_sync_orchestrion.txt node --import ./sync-hooks-orchestrion.mjs runtime.mjs
 /usr/bin/time -v -o time_sync_orchestrion_tracing.txt node --import ./sync-hooks-orchestrion-tracing.mjs runtime.mjs
-/usr/bin/time -v -o time_sync_regex.txt node --import ./sync-hooks-regex.mjs runtime.mjs
-/usr/bin/time -v -o time_async_babel_one_file.txt node --import ./async-hooks-babel-one-file.mjs runtime.mjs
-/usr/bin/time -v -o time_async_babel.txt node --import ./register-async-hooks-babel.mjs runtime.mjs
-/usr/bin/time -v -o time_async_oxc.txt node --import ./register-async-hooks-oxc.mjs runtime.mjs
-/usr/bin/time -v -o time_async_regex.txt node --import ./register-async-hooks-regex.mjs runtime.mjs
 
 node update-bench-table.mjs
 node make-chart.mjs