+++ hooks/register-async-hooks-oxc.mjs
@@ -1,2 +1,0 @@
-import { register } from 'node:module'
-register('./async-hooks-oxc.mjs', import.meta.url)