+++ hooks/runtime.mjs
@@ -1,4 +1,3 @@
-import './wrap.mjs'
 import { handler } from './handler.mjs'
 
 const res = await handler({ foo: 'bar' }, {})