+++ hooks/sourcemap-ts-demo/wrap-noop.mjs
@@ -1,3 +1,0 @@
-globalThis.WrapAwsLambda = function (orig_handler) {
-  return async (event, context) => orig_handler(event, context)
-}